Realtionship feature

+++ src/main/java/com/cjk/stackcast/services/VideoService.java
@@ -20,20 +20,19 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.util.Date;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class VideoService {
 
-    @Autowired
     private AwsS3Configuration config;
-
-    @Autowired
     private VideoRepository repo;
 
     @Autowired
-    private UserService userService;
+    public VideoService(VideoRepository repo, AwsS3Configuration config){
+        this.repo = repo;
+        this.config = config;
+    }
 
     public Optional<Video> show(Long videoId){
             return repo.findById(videoId);
@@ -44,37 +43,26 @@ public class VideoService {
     }
 
     public Iterable<Video> showAllUserVids(Long userId){
-        return repo.findByUser_Id(userId);
+        return repo.findAllByUserId(userId);
     }
+
 
     public Video createVideo(Video basicVideo){
         return repo.save(basicVideo);
     }
 
-    public Video saveVideo(String videoName, MultipartFile multipartFile) throws Exception{
-        String endPointUrl = "https://zip-code-video-app.s3.amazonaws.com";
-        File file = convertMultiPartFile(multipartFile);
-        Video video = new Video(videoName,multipartFile.getContentType());
-        String fileName = generateFileName(file.getName());
-        String fileUrl = endPointUrl + "/" + fileName;
-        video.setVideoPath(fileUrl);
-
-        if(uploadFile(file,fileName).isSuccessful()){
-           return createVideo(video);
+    public boolean delete(Long videoId){
+        String originalVideoKey = repo.getOne(videoId).getOriginalVideoKey();
+        if(deleteFile(originalVideoKey).isSuccessful()){
+            repo.deleteById(videoId);
+            return true;
         } else
-            return null;
-    }
-
-    public boolean delete(Long videoId) throws Exception {
-        //TODO resolve delete from s3 bucket per "key"(filename)
-        repo.deleteById(videoId);
-        return true;
+            return false;
     }
 
     public Video setUser(Long videoId, Long userId){
         Video video = repo.getOne(videoId);
-        User user = userService.getUser(userId);
-        video.setUser(user);
+        video.setUserId(userId);
         return repo.save(video);
     }
 
@@ -94,6 +82,18 @@ public class VideoService {
         Video video = repo.getOne(videoId);
         video.setVideoPath(newPath);
         return repo.save(video);
+    }
+
+    public Video uploadVideo(String videoName, MultipartFile multipartFile) throws Exception{
+        File file = convertMultiPartFile(multipartFile);
+        String fileName = generateFileName(file.getName());
+        Video video = new Video(videoName,multipartFile.getContentType(),fileName);
+        String fileUrl = config.getEndPointUrl() + "/" + fileName;
+        video.setVideoPath(fileUrl);
+        if(uploadFile(file,fileName).isSuccessful()){
+            return createVideo(video);
+        } else
+            return null;
     }
 
     public File convertMultiPartFile(MultipartFile file) throws IOException {
@@ -116,18 +116,19 @@ public class VideoService {
         return config.generateS3Client().putObject(putObjectRequest, RequestBody.fromFile(file)).sdkHttpResponse();
     }
 
-    public DeleteObjectResponse deleteFile(String fileName){
-        DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
-                .bucket(config.getBucket()).key(fileName).build();
-        return config.generateS3Client().deleteObject(deleteObjectRequest);
+    public void verifyFileType(String videoName, MultipartFile multipartFile) throws Exception{
+        ArrayList<String> validFileTypes = new ArrayList<>(Arrays.asList("video/mp4","video/mov"));
+        if(validFileTypes.contains(multipartFile.getContentType())){
+            uploadVideo(videoName, multipartFile);
+        } else {
+            throw new InputMismatchException("Invalid file type. Only video files are allowed");
+        }
     }
 
-
-
-
-
-
-
-
+    public SdkHttpResponse deleteFile(String fileName){
+        DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
+                .bucket(config.getBucket()).key(fileName).build();
+        return config.generateS3Client().deleteObject(deleteObjectRequest).sdkHttpResponse();
+    }
 
 }
